Create updates.cs
The `key` variable is declared twice, once at the beginning of the `EncryptDir` method and once inside the `foreach` loop. The value of the second statement is being used to encrypt the file, but this variable is not being passed to the `EncryptFile` method.

There is a line break and non-ASCII characters in the `key` string. This can cause problems when converting the string to ASCII bytes using `ASCIIEncoding.ASCII.GetBytes(sKey)`.

The code is using DES encryption algorithm which is currently considered insecure. It is recommended to use more secure algorithms such as AES.

Here is a modified version of the code that addresses these issues: